pyproject.toml
[tool.poetry]
name = "QQRobot"
version = "0.1.0"
description = "QQRobot"
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "^2.0.0-beta.1"
PyMySQL = "^1.0.2"
httpx = "^0.23.0"
nonebot-adapter-onebot = "^2.1.1"

[tool.poetry.dev-dependencies]
nb-cli = "^0.6.0"

[tool.nonebot]
plugin_dirs = ["src/plugins"]

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
